touch/novatek: fix reporting 0 battery level
The controller FW sends coordinates with battery level 0 at the early
stage of pairing. It's because the FW gets a correct battery level
after the pairing is done. If a stylus is used for very short time then
the FW has no time to get the correct battery level from the stylus.
So the driver should ignore those reports with battery level 0.

This patch ignores the zero battery level.

Test: The battery level never goes to 0 with a very short stroke.
Bug: 240261329
Signed-off-by: Hyungwoo Yang <hyungwooyang@google.com>
Change-Id: I40858ac7d6a73b5b1c16c78e91b8d212d1b605d0