touch/novatek: stylus info handling on normal fairing
Sometimes, even with the only one stylus, we see normal pairing
happening instead of fast pairing. This removes the stylus
information maintained in the driver and the UI platform cannot
shows the stylus information until the driver collects the same
information again.

This patch, when there's normal pairing, compares new HASH ID with the
one from the previous stylus. The driver removes the stylus information
only when the HASH ID is changed.

Bug: 259261326

Change-Id: I1cda091a3455d8e7adcdecfeeded5de69bbece5d
Signed-off-by: Hyungwoo Yang <hyungwooyang@google.com>