vh: sched: Remove sugov util from avg_rt
Although we zero sugov tasks util_avg, but GKI already adds it to
rq->avg_rt->util_avg. We need to make sure to undo that which
unfortunately means we have to calculate the util_avg of sugov task to
subtract the value then zero it.

Add trace_pelt_rq_rt() since we update this value and it helps to debug
rq->avg_rt values.

Bug: 394668067
Signed-off-by: Qais Yousef <qyousef@google.com>
Change-Id: I6bd57318d2441229ec187681622be64994eda880