uwb: qm35: Don't accidentally start when we were off
The qm35 driver has several instances of code paths that look like
this:

  qm35_hsspi_stop();
  /* do something */
  qm35_hsspi_start();

The issue is that some of these code paths run in cases where the QM35
might have already been stopped. Annotating the above code to show
what happens if it's run when we started out in the "stopped" case:

  qm35_hsspi_stop();  // <-- NOOP because we were already stopped
  /* do something */
  qm35_hsspi_start(); // <-- Now we'll start

It should be noted that qm35_hsspi_start() doesn't actually power on
the qm35. To do that, you need to turn on the regulators (see
QM35_CTRL_POWER). That means that running the problematic sequence
will make some parts of the code think things are started but they
won't really be.

One especially troubling case of the above sequence causing problems
is during suspend/resume. If we started off stopped/powered off
(because UWB is off in the system preferences) then going through the
suspend/resume cycle will leave the regulators off but put the driver
in "running" state.

When the driver is in the "running" state without regulators, it tends
to print lots of messages like:
  qm35 spi18.0: timed out waiting for ss_ready(0)
  qm35 spi18.0: timed out waiting for ss_ready(0)
  qm35 spi18.0: timed out waiting for ss_ready(0)
  qm35 spi18.0: timed out waiting for ss_ready(0)
  qm35 spi18.0: timed out waiting for ss_ready(0)
  qm35 spi18.0: FW not ready (flags 0x0)
  ...
  Max successive errors 6 reached, likely entered ROM code...
  No response from FW after multiple resets. Stopping HSSPI.

Specifically, these errors tend to get triggered from subseuent
suspend/resume calls which are trying to properly put QM35 into
suspend since the driver is confused about the state.

A related problem is often triggered by all of the above confusion
that is reported in the kernel as something like this (actual IRQ
number and offsets can vary):

  Unbalanced enable for IRQ 528
  ...
  enable_irq+0x98/0xf4
  qm35_pm_resume+0x5c/0x184 [qm35]
  ...

That specifically happens because near where the driver printed the
error message "No response from FW after multiple resets. Stopping
HSSPI." it jams the driver into `HSSPI_STOPPED` state without
disabling interrupts. That confuses subsequent suspend/resume
calls. Presumably the driver should handle the error more elegantly,
but for now we'll simply avoid the problem by preventing the bad state
to start with.

To fix these problems, audit all code calling qm35_hsspi_stop() and
only call qm35_hsspi_start() if we were in the "running" state to
begin with. To handle suspend/resume time we'll create a new state
"HSSPI_SUSPENDED" that represents that QM35 was stopped for the
purposes of suspend and should be started again at resume time.

NOTE: Much of the state handling in QM35 could probably benefit from
thinking more about how the locking is supposed to work and making
sure all the paths are safe. This patch doesn't attempt any cleanup
around locking but is not expected to make things any worse.

Bug: 391402684
Bug: 390854151
Bug: 391464939
Test: Suspend/resume and no more qm35 errors

Change-Id: I19585dfd50ca24a69cb9e82623f7f1dff602f219
Signed-off-by: Douglas Anderson <dianders@google.com>